label-reader/modules/web_app_api/requirements.txt
# Runtime
azure-iot-device==2.10.0
tornado==6.1
opencensus-ext-azure==1.1.1
opencensus-ext-logging==0.1.1
opencensus-ext-requests==0.7.6
msgpack==1.0.3
azure-storage-blob==12.11.0
# Devlopment
nose2==0.11.0
black==22.1.0